pyproject.toml
[project]
name = "flare"
version = "0.1.0"
description = ""
authors = [
    {name = "JuiHsuanLee0303",email = "juihsuanlee0303@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi (>=0.115.12,<0.116.0)",
    "uvicorn (>=0.34.2,<0.35.0)",
    "dotenv (>=0.9.9,<0.10.0)",
    "requests (>=2.32.3,<3.0.0)",
    "pypdf (>=5.5.0,<6.0.0)",
    "chardet (>=5.2.0,<6.0.0)",
    "python-docx (>=1.1.2,<2.0.0)"
]

[tool.poetry]
packages = [{include = "flare", from = "src"}]


[tool.poetry.group.rag.dependencies]
qdrant-client = "^1.14.2"


[tool.poetry.group.dev.dependencies]
pytest = "^8.3.5"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
